+++ IPython/core/interactiveshell.py
@@ -106,68 +106,13 @@ class ProvisionalWarning(DeprecationWarning):
 # Await Helpers
 #-----------------------------------------------------------------------------
 
-from textwrap import indent, dedent
-import ast
 
-
-def _asyncio_runner(function, user_ns):
-    import asyncio
-    return asyncio.get_event_loop().run_until_complete(function(**user_ns))
-
-def _curio_runner(function, user_ns):
-    import curio
-    return curio.run(function(**user_ns))
-
-def _trio_runner(function, user_ns):
-    import trio
-    async def loc(fun, user_ns):
-        """
-        We need the dummy no-op async def to protect from
-        trio's internal. See https://github.com/python-trio/trio/issues/89
-        """
-        return await fun(**user_ns)
-    return trio.run(loc, function, user_ns)
-
-
-def _asyncify(code):
-    """wrap code in async def definition.
-
-    And setup a bit of context to run it later.
-
-    The following names are part of the API:
-
-     - ``user_ns`` is used as the name for the namespace to run code in, both
-     _in_ and _out_.
-     - ``loop_runner`` is use to send the loop_runner _in_
-     - ``last_expression`` to get the last expression _out_
-    """
-    return dedent("""
-        async def phony():
-        {usercode}
-            return locals()
-        interm = loop_runner(phony, user_ns)
-        user_ns, last_expr =  interm
-        """).format(usercode=indent(code,' '*4))
-
-
-def _should_be_async(cell:str) -> bool:
-    """Detect if a block of code need to be wrapped in an `async def`
-
-    Attempt to parse the block of code, it it compile we're fine.
-    Otherwise we  wrap if and try to compile.
-
-    If it works, assume it should be async. Otherwise Return False.
-    """
-    try:
-        ast.parse(cell)
-        return False
-    except SyntaxError:
-        try:
-            ast.parse(_asyncify(cell))
-        except SyntaxError:
-            return False
-        return True
-    return False
+if sys.version_info > (3,5):
+    from .async_helpers import (_asyncio_runner, _curio_runner, _trio_runner,
+                                _should_be_async, _asyncify)
+else :
+    _asyncio_runner = _curio_runner = _trio_runner = None
+    _should_be_async = lambda x : False
 
 
 def _ast_asyncify(cell:str, localnames) -> ast.Module:
@@ -1497,8 +1442,12 @@ class InteractiveShell(SingletonConfigurable):
                            ]
 
         # initialize results to 'null'
-        found = False; obj = None;  ospace = None;
-        ismagic = False; isalias = False; parent = None
+        obj = None
+        parent = None
+        ospace = None
+        found = False
+        ismagic = False
+        isalias = False
 
         # Look for the given name by splitting it in parts.  If the head is
         # found, then we look for all the remaining parts as members, and only
@@ -1511,7 +1460,6 @@ class InteractiveShell(SingletonConfigurable):
             except KeyError:
                 continue
             else:
-                #print 'oname_rest:', oname_rest  # dbg
                 for idx, part in enumerate(oname_rest):
                     try:
                         parent = obj
@@ -3125,7 +3073,7 @@ class InteractiveShell(SingletonConfigurable):
                 self.pylab_gui_select = gui
             # Otherwise if they are different
             elif gui != self.pylab_gui_select:
-                print ('Warning: Cannot change to a different GUI toolkit: %s.'
+                print('Warning: Cannot change to a different GUI toolkit: %s.'
                         ' Using %s instead.' % (gui, self.pylab_gui_select))
                 gui, backend = pt.find_gui_and_backend(self.pylab_gui_select)
